+++ frontends/web/src/contexts/AppProvider.tsx
@@ -1,5 +1,5 @@
 /**
- * Copyright 2023 Shift Crypto AG
+ * Copyright 2023-2024 Shift Crypto AG
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,12 +17,17 @@
 import { ReactNode, useEffect, useState } from 'react';
 import { getConfig, setConfig } from '../utils/config';
 import { AppContext, TSidebarStatus } from './AppContext';
+import { useLoad } from '../hooks/api';
+import { useDefault } from '../hooks/default';
+import { getNativeLocale } from '../api/nativelocale';
+import { i18nextFormat } from '../i18n/utils';
 
 type TProps = {
     children: ReactNode;
 }
 
 export const AppProvider = ({ children }: TProps) => {
+  const nativeLocale = i18nextFormat(useDefault(useLoad(getNativeLocale), 'de-CH'));
   const [guideShown, setGuideShown] = useState(false);
   const [guideExists, setGuideExists] = useState(false);
   const [hideAmounts, setHideAmounts] = useState(false);
@@ -66,6 +71,7 @@ export const AppProvider = ({ children }: TProps) => {
         guideShown,
         guideExists,
         hideAmounts,
+        nativeLocale,
         sidebarStatus,
         setActiveSidebar,
         setGuideShown,